Правки миграций под mssql

+++ migrations/20200229233751-categories_bck.sql
@@ -1,6 +1,6 @@
 
 -- +migrate Up
-	CREATE TABLE categories_bck (id integer(11) NOT NULL, description varchar(255));
+	CREATE TABLE categories_bck (id int NOT NULL, description varchar(255));
 	
 -- +migrate Down
 	DROP TABLE categories_bck;
